requirements/base.txt
# Paquetes compartidos (Django, DRF, CORS, etc.)

#   Instrucciones para instalación
    # pip install -r requirements/base.txt

Django==4.1.7
djangorestframework==3.14.0
django-cors-headers==3.14.0

# Documentación y utilidades
coreapi==2.3.3
coreschema==0.0.4
uritemplate==4.1.1
itypes==1.2.0

# Dependencias generales de Python
pytz==2022.7.1
sqlparse==0.4.3
asgiref==3.6.0
requests==2.28.2
charset-normalizer==3.1.0
idna==3.4
urllib3==1.26.15
certifi==2022.12.7
